Extract provider-neutral review-loop prompt evidence
